ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
python-multipart==0.0.6
openai
pdf2image
pillow
